Support `no_std`
Resolves #43
